app/requirements.txt
Brotli==1.0.9
certifi==2020.6.20
click==7.1.2
cycler==0.10.0
dash==1.16.0
dash-core-components==1.12.0
dash-html-components==1.1.1
dash-renderer==1.8.0
dash-table==4.10.1
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.1
nltk==3.5
numpy==1.19.1
pandas==1.1.2
Pillow==7.2.0
plotly==4.10.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
retrying==1.3.3
six==1.15.0
tqdm==4.48.2
Werkzeug==1.0.1
wordcloud==1.8.0
google-cloud-storage==1.31.2
Flask-Caching==1.9.0
ipywidgets==7.5.1